# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Marketplace AI API",
    "version": "1.0.0",
    "description": "API documentation for the Marketplace AI platform"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for webhook authentication"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Error status"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            },
            "description": "Validation errors"
          }
        }
      },
...
      "ComplexResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the operation was successful"
          },
          "message": {
            "type": "string",
            "description": "Response message"
          },
          "data": {
            "type": "object",
            "description": "Response data"
          },
          "metadata": {
            "type": "object",
            "description": "Additional response metadata",
            "properties": {
              "processingTime": {
                "type": "number"
              },
              "serverTime": {
                "type": "string",
                "format": "date-time"
              },
              "pagination": {
                "$ref": "#/components/schemas/Pagination"
              },
              "filters": {
                "type": "object"
              },
              "sorting": {
                "type": "object"
              }
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                },
                "details": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/admin/listings": {
      "get": {
        "summary": "Get list of listings",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of listings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplexResponse"
                },
                "examples": {
                  "ListResponse": {
                    "$ref": "#/components/examples/ResponseExamples/ListingCreated"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthorizationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthorizationError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/listings/{listingId}": {
      "patch": {
        "summary": "Update a listing by ID",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "listingId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Listing ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Listing"
              },
              "examples": {
                "ListingExample": {
                  "$ref": "#/components/examples/ListingExample"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Listing updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "ListingExample": {
                    "$ref": "#/components/examples/ListingExample"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthorizationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthorizationError"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Listing not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "NotFoundError": {
                    "$ref": "#/components/examples/ErrorExamples/NotFoundError"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a listing by ID",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "listingId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Listing ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Listing deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "SuccessResponse": {
                    "value": {
                      "message": "Listing deleted",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthorizationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthorizationError"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Listing not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "NotFoundError": {
                    "$ref": "#/components/examples/ErrorExamples/NotFoundError"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/listings": {
      "get": {
        "summary": "Get all listings with filtering, search, and pagination",
        "tags": [
          "Listings"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "sale",
                "auction",
                "rental"
              ]
            },
            "description": "Filter by listing type"
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Filter by category"
          },
          {
            "in": "query",
            "name": "minPrice",
            "schema": {
              "type": "number"
            },
            "description": "Minimum price"
          },
          {
            "in": "query",
            "name": "maxPrice",
            "schema": {
              "type": "number"
            },
            "description": "Maximum price"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Keyword search"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "enum": [
                "price",
                "createdAt",
                "relevance"
              ]
            },
            "description": "Sort by field"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of listings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplexResponse"
                },
                "examples": {
                  "SearchResults": {
                    "$ref": "#/components/examples/ComplexScenarios/SearchResults"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "ValidationError": {
                    "$ref": "#/components/examples/ErrorExamples/ValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new listing",
        "tags": [
          "Listings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Listing"
              },
              "examples": {
                "CreateListing": {
                  "$ref": "#/components/examples/RequestExamples/CreateListing"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Listing created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "ListingCreated": {
                    "$ref": "#/components/examples/ResponseExamples/ListingCreated"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "ValidationError": {
                    "$ref": "#/components/examples/ErrorExamples/ValidationError"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/listings/{listingId}": {
      "get": {
        "summary": "Get listing by ID",
        "tags": [
          "Listings"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "listingId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Listing ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Listing details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "ListingExample": {
                    "$ref": "#/components/examples/ListingExample"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Listing not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "NotFoundError": {
                    "$ref": "#/components/examples/ErrorExamples/NotFoundError"
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update listing",
        "tags": [
          "Listings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "listingId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Listing ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Listing"
              },
              "examples": {
                "UpdateListing": {
                  "$ref": "#/components/examples/RequestExamples/UpdateListing"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Listing updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "ListingUpdated": {
                    "$ref": "#/components/examples/ResponseExamples/ListingUpdated"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "ValidationError": {
                    "$ref": "#/components/examples/ErrorExamples/ValidationError"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Listing not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "NotFoundError": {
                    "$ref": "#/components/examples/ErrorExamples/NotFoundError"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete listing",
        "tags": [
          "Listings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "listingId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Listing ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Listing deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "SuccessResponse": {
                    "value": {
                      "message": "Listing deleted successfully",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Listing not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "NotFoundError": {
                    "$ref": "#/components/examples/ErrorExamples/NotFoundError"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orders/listings/{listingId}": {
      "post": {
        "summary": "Create a new order for a listing",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "listingId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Listing ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "shippingAddress": {
                    "$ref": "#/components/schemas/Order/properties/shippingAddress"
                  },
                  "paymentMethodId": {
                    "type": "string",
                    "description": "Payment method ID"
                  }
                }
              },
              "examples": {
                "CreateOrder": {
                  "value": {
                    "shippingAddress": {
                      "street": "123 Main St",
                      "city": "New York",
                      "state": "NY",
                      "country": "USA",
                      "zipCode": "10001"
                    },
                    "paymentMethodId": "pm_1Hh1Y2A3B4C5D6E7F8G9H0"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "examples": {
                  "OrderExample": {
                    "$ref": "#/components/examples/OrderExample"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "ValidationError": {
                    "$ref": "#/components/examples/ErrorExamples/ValidationError"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "AuthenticationError": {
                    "$ref": "#/components/examples/ErrorExamples/AuthenticationError"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}